--- src/main/java/com/wd40/satisfactorymanager/dto/change/ChangeOp.java
@@ -1,7 +1,0 @@
-package com.wd40.satisfactorymanager.dto.change;
-
-public enum ChangeOp {
-  ADD,
-  UPDATE,
-  DELETE
-}
