+++ src/main/java/com/wd40/satisfactorymanager/dto/request/change/ChangeOp.java
@@ -1,0 +1,7 @@
+package com.wd40.satisfactorymanager.dto.request.change;
+
+public enum ChangeOp {
+  ADD,
+  UPDATE,
+  DELETE
+}
