+++ src/main/java/com/wd40/satisfactorymanager/util/FactoryMapper.java
@@ -1,6 +1,6 @@
 package com.wd40.satisfactorymanager.util;
 
-import com.wd40.satisfactorymanager.dto.FactoryDto;
+import com.wd40.satisfactorymanager.dto.response.FactoryDto;
 import com.wd40.satisfactorymanager.model.Factory;
 import org.mapstruct.Mapper;
 import org.mapstruct.factory.Mappers;
